requirements.txt
# FileSage 2.0 Dependencies
# Core functionality
PyYAML>=6.0.2

# Optional: For future enhancements
# click>=8.0.0          # CLI improvements
# mermaid-py>=0.12.0    # Visual diagrams (future)
# jinja2>=3.1.0         # Template system (future)
